+++ src/com/osudroid/beatmaplisting/BeatmapMirrorModels.kt
@@ -1,4 +1,4 @@
-package com.reco1l.osu.beatmaplisting
+package com.osudroid.beatmaplisting
 
 import okhttp3.HttpUrl
 import ru.nsu.ccfit.zuev.osu.RankedStatus
