+++ src/com/osudroid/beatmaplisting/mirrors/Catboy.kt
@@ -1,21 +1,35 @@
-package com.reco1l.osu.beatmaplisting.mirrors
+package com.osudroid.beatmaplisting.mirrors
 
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorDownloadRequestModel
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorPreviewRequestModel
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorSearchRequestModel
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorSearchRequestModel.*
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorSearchRequestModel.OrderType.*
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.*
-import com.reco1l.osu.beatmaplisting.BeatmapMirrorSearchResponseModel
-import com.reco1l.osu.beatmaplisting.BeatmapModel
-import com.reco1l.osu.beatmaplisting.BeatmapSetModel
+import com.osudroid.beatmaplisting.BeatmapMirrorDownloadRequestModel
+import com.osudroid.beatmaplisting.BeatmapMirrorPreviewRequestModel
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.*
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.OrderType.Ascending
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.OrderType.Descending
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.Artist
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.BPM
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.DifficultyRating
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.FavouriteCount
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.HitLength
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.LastUpdated
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.PassCount
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.PlayCount
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.RankedDate
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.SubmittedDate
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.Title
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchRequestModel.SortType.TotalLength
+import com.osudroid.beatmaplisting.BeatmapMirrorSearchResponseModel
+import com.osudroid.beatmaplisting.BeatmapModel
+import com.osudroid.beatmaplisting.BeatmapSetModel
 import okhttp3.HttpUrl
 import okhttp3.HttpUrl.Companion.toHttpUrl
 import org.json.JSONArray
 import ru.nsu.ccfit.zuev.osu.RankedStatus
 
+// API reference: https://dev.catboy.best/docs
 
-class OsuDirectSearchRequestModel : BeatmapMirrorSearchRequestModel {
+
+class CatboySearchRequestModel : BeatmapMirrorSearchRequestModel {
 
     private fun SortType.parseToApi() = when (this) {
         Title -> "title"
@@ -39,7 +53,7 @@ class OsuDirectSearchRequestModel : BeatmapMirrorSearchRequestModel {
 
     override fun invoke(query: String, offset: Int, limit: Int, sort: SortType, order: OrderType, status: RankedStatus?): HttpUrl {
 
-        return "https://osu.direct/api/v2/search".toHttpUrl()
+        return "https://catboy.best/api/v2/search".toHttpUrl()
             .newBuilder()
             .addQueryParameter("sort", "${sort.parseToApi()}:${order.parseToApi()}")
             .apply {
@@ -49,18 +63,19 @@ class OsuDirectSearchRequestModel : BeatmapMirrorSearchRequestModel {
             }
             .addQueryParameter("mode", "0")
             .addQueryParameter("query", query)
+            .addQueryParameter("limit", limit.toString())
             .addQueryParameter("offset", offset.toString())
-            .addQueryParameter("amount", limit.toString())
             .build()
     }
 }
 
-class OsuDirectSearchResponseModel : BeatmapMirrorSearchResponseModel {
+
+class CatboySearchResponseModel : BeatmapMirrorSearchResponseModel {
     override fun invoke(response: Any): MutableList<BeatmapSetModel> {
         response as JSONArray
 
-        return MutableList(response.length()) { index ->
-            val json = response.getJSONObject(index)
+        return MutableList(response.length()) { i ->
+            val json = response.getJSONObject(i)
 
             BeatmapSetModel(
                 id = json.getLong("id"),
@@ -70,7 +85,7 @@ class OsuDirectSearchResponseModel : BeatmapMirrorSearchResponseModel {
                 artistUnicode = json.getString("artist_unicode"),
                 status = RankedStatus.valueOf(json.getInt("ranked")),
                 creator = json.getString("creator"),
-                thumbnail = json.optJSONObject("covers")?.optString("card"),
+                thumbnail = "https://assets.ppy.sh/beatmaps/${json.getLong("id")}/covers/card.jpg",
                 beatmaps = json.getJSONArray("beatmaps").let {
 
                     MutableList(it.length()) { i ->
@@ -99,14 +114,16 @@ class OsuDirectSearchResponseModel : BeatmapMirrorSearchResponseModel {
     }
 }
 
-class OsuDirectDownloadRequestModel : BeatmapMirrorDownloadRequestModel {
+
+class CatboyDownloadRequestModel : BeatmapMirrorDownloadRequestModel {
     override fun invoke(beatmapSetId: Long): HttpUrl {
-        return "https://osu.direct/api/d/$beatmapSetId".toHttpUrl()
+        return "https://catboy.best/d/$beatmapSetId".toHttpUrl()
     }
 }
 
-class OsuDirectPreviewRequestModel : BeatmapMirrorPreviewRequestModel {
+
+class CatboyPreviewRequestModel : BeatmapMirrorPreviewRequestModel {
     override fun invoke(beatmapId: Long): HttpUrl {
-        return "https://osu.direct/api/media/preview/$beatmapId".toHttpUrl()
+        return "https://catboy.best/preview/audio/$beatmapId".toHttpUrl()
     }
 }
