+++ src/com/reco1l/osu/ui/entity/BeatmapButton.kt
@@ -5,8 +5,8 @@ import com.osudroid.ibancho.data.PlayerStatus.Ready
 import com.osudroid.ibancho.data.RoomBeatmap
 import com.reco1l.osu.async
 import com.reco1l.osu.multiplayer.Multiplayer
-import com.reco1l.osu.beatmaplisting.BeatmapDownloader
-import com.reco1l.osu.beatmaplisting.BeatmapListing
+import com.osudroid.beatmaplisting.BeatmapDownloader
+import com.osudroid.beatmaplisting.BeatmapListing
 import com.reco1l.osu.multiplayer.RoomScene
 import org.anddev.andengine.entity.sprite.Sprite
 import org.anddev.andengine.entity.text.ChangeableText
